requirements.txt
asgiref==3.2.10
Django==3.0.8
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==8.1.0
python-decouple==3.4
pytz==2020.1
sqlparse==0.3.1
tesseract==0.1.3
